+++ Plutarch/Prelude.hs
@@ -1,3 +1,5 @@
+{-# OPTIONS_GHC -Wno-deprecations #-}
+
 module Plutarch.Prelude (
   -- * Boolean
   PBool (..),