+++ plutarch-ledger-api/CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## Unreleased
+
+### Added
+
+* `SOP.Generic` derivations for `PDatum`, `PRedeemer`.
+
+### Changed
+
+* Replaced all internal uses of `pfixHoisted` with `pfix` for improved
+  performance, at the cost of a negligible increase in script size in some
+  cases.
+
+### Removed
+
+* `DerivePlutusType` instance for `PPosixTime`, `PDatum`, `PRedeemer` as 
+  `DerivePlutusType` is deprecated now. 
+
 ## 3.5.0 -- 07/11/2025
 
 ### Added